fix issues, 1.0.0.6

+++ CataclysmMod.IL.FungalClumpDamage.cs
@@ -1,5 +1,6 @@
 ﻿using Mono.Cecil.Cil;
 using MonoMod.Cil;
+using Terraria;
 using Terraria.ModLoader;
 
 namespace CataclysmMod
@@ -8,24 +9,26 @@ namespace CataclysmMod
     {
         private void RemoveSummonDamageBonus(ILContext il)
         {
-            ILCursor c = new ILCursor(il);
+			if (!Main.dedServ) {
+				ILCursor c = new ILCursor(il);
 
-            if (!c.TryGotoNext(i => i.MatchLdcR4(10)))
-            {
-                Logger.Warn("[IL] Unable to match ldc.r4 \"10\"!");
-                return;
-            }
+				if (!c.TryGotoNext(i => i.MatchLdcR4(10)))
+				{
+					Logger.Warn("[IL] Unable to match ldc.r4 \"10\"!");
+					return;
+				}
+				
+				c.Index++;
 
-            c.Index++;
+				c.Emit(OpCodes.Pop);
+				c.Emit(OpCodes.Ldc_I4, 10);
 
-            c.Emit(OpCodes.Pop);
-            c.Emit(OpCodes.Ldc_I4, 10);
+				c.Index++;
 
-            c.Index++;
+				c.RemoveRange(4);
 
-            c.RemoveRange(4);
-
-            Logger.Info("[IL] Finished patching!");
+				Logger.Info("[IL] Finished patching!");
+			}
         }
     }
 }
